+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2020/06/09
+
+1. 暂存问题：
+    - 昨日将数据库所有内容改为小写后原 Article 部分的 VO 出现了部分字段不显示的问题
+    - 暂未完成在 services 层写出 classify 应返回的 classifyList
+    - 暂未完成剩余的 classify 部分的增、删、改、查
+
+---
+
 ## 2020/06/08
 
 1. 修正了 getArticleById 的 router 层中没有以正常逻辑进入 VO